fastapi_metagpt_integration/requirements.txt
fastapi==   0.110.0
uvicorn==0.15.0
# metagpt==0.7.4
python-dotenv==1.0.0
httpx==0.27.0
fastjsonschema==2.19.1
gitdb==4.0.11
gitignore_parser==0.1.9
GitPython==3.1.40
loguru==0.6.0
multiprocess==0.70.16
pydantic==2.5.3
pytest==8.0.1